+++ changelog.md
@@ -38,6 +38,8 @@ release. Some specific details may be omitted for brevity.
 
 ### No longer broken
 
+- Fixed: Inline formula rendering now correctly renders lists on the same line.
+- Fixed: Markdown rendered with `md()` in formula properties within the bases `List` view now correctly renders on the same line.
 - Fixed: The global scope version of `then()` now behaves as expected when only one parameter is provided.
 - Fixed: There is no longer an error in the dev console on vault startup if FF is enabled.